Addressing Zach's github comments. These involved renameing the class from SafeWrite to ConditionalFileUpdater.
