+++ nrpy/infrastructures/CarpetX/interface_ccl.py
@@ -9,7 +9,7 @@ Author: Zachariah B. Etienne
 from typing import List
 from pathlib import Path
 import nrpy.grid as gri
-from nrpy.helpers.safewrite import SafeWrite
+from nrpy.helpers.conditional_file_updater import ConditionalFileUpdater
 
 
 def construct_interface_ccl(
@@ -152,5 +152,5 @@ public:
 """
     output_Path = Path(project_dir) / thorn_name
     output_Path.mkdir(parents=True, exist_ok=True)
-    with SafeWrite(output_Path / "interface.ccl", encoding="utf-8") as file:
+    with ConditionalFileUpdater(output_Path / "interface.ccl", encoding="utf-8") as file:
         file.write(outstr)
