+++ nrpy/infrastructures/ETLegacy/make_code_defn.py
@@ -9,7 +9,7 @@ from typing import List
 from pathlib import Path
 
 import nrpy.c_function as cfc
-from nrpy.helpers.safewrite import SafeWrite
+from nrpy.helpers.conditional_file_updater import ConditionalFileUpdater
 
 
 def output_CFunctions_and_construct_make_code_defn(
@@ -43,14 +43,14 @@ def output_CFunctions_and_construct_make_code_defn(
     make_code_defn_file = src_Path / "make.code.defn"
 
     # Open and write to the make.code.defn file
-    with SafeWrite(make_code_defn_file) as make_code_defn:
+    with ConditionalFileUpdater(make_code_defn_file) as make_code_defn:
         make_code_defn.write(f"# make.code.defn file for thorn {thorn_name}\n\n")
         make_code_defn.write("# Source files that need to be compiled:\n")
         make_code_defn.write("SRCS = \\\n")
 
         # Iterate through sorted list of CFunctions and write each to the make.code.defn file
         for i, CFunction in enumerate(make_code_defn_list_of_CFunctions):
-            with SafeWrite(src_Path / f"{CFunction.name}.c", encoding="utf-8") as file:
+            with ConditionalFileUpdater(src_Path / f"{CFunction.name}.c", encoding="utf-8") as file:
                 file.write(CFunction.full_function)
 
             # If it's not the last iteration, append a backslash:
